+++ R/cluster_determination.R
@@ -588,16 +588,18 @@ BuildClusterTree <- function(
   }
   if (! is.null(x = SNN.use)) {
     num.clusters <- length(x = ident.names)
-    data.dist = matrix(data = 0, nrow = num.clusters, ncol = num.clusters)
+    data.dist <- matrix(data = 0, nrow = num.clusters, ncol = num.clusters)
+    rownames(data.dist) <- ident.names
+    colnames(data.dist) <- ident.names
     for (i in 1:(num.clusters - 1)) {
       for (j in (i + 1):num.clusters) {
         subSNN <- SNN.use[
           match(
-            x = WhichCells(object = object, ident = i),
+            x = WhichCells(object = object, ident = ident.names[i]),
             table = colnames(x = SNN.use)
           ), # Row
           match(
-            x = WhichCells(object = object, ident = j),
+            x = WhichCells(object = object, ident = ident.names[j]),
             table = rownames(x = SNN.use)
           ) # Column
           ]
@@ -605,7 +607,7 @@ BuildClusterTree <- function(
         if (is.na(x = d)) {
           data.dist[i, j] <- 0
         } else {
-          data.dist[i, j] = d
+          data.dist[i, j] <- d
         }
       }
     }