+++ src/snn.h
@@ -8,13 +8,14 @@
 #include <unordered_map>
 #include <fstream>
 #include <string>
+#include <iomanip>
 
 using namespace Rcpp;
 
 //----------------------------------------------------
 Eigen::SparseMatrix<double> ComputeSNN(Eigen::MatrixXd nn_large, Eigen::MatrixXd nn_ranked, double prune, bool display_progress);
 void WriteEdgeFile(Eigen::SparseMatrix<double> snn, String filename, bool display_progress);
-Eigen::SparseMatrix<double> DirectToFile(Eigen::MatrixXd nn_large, Eigen::MatrixXd nn_ranked, double prune, bool display_progress, String filename);
+Eigen::SparseMatrix<double> DirectSNNToFile(Eigen::MatrixXd nn_large, Eigen::MatrixXd nn_ranked, double prune, bool display_progress, String filename);
 //----------------------------------------------------
 
 #endif//SNN